Refactor generator functions to their own group
Also renamed "Collector" group to "ResourceTypes".
