+++ R.swift/ResourceTypes/AssetFolder.swift
