+++ R.swift/ResourceTypes/Image.swift
